requirements.txt
img2pdf
pillow
opencv-python
pyautogui
pyqt5
tkinter